requirements.txt
# === Core Framework ===
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1

# === Database & Auth ===
firebase-admin==6.5.0
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2

# === LLM & AI ===
google-generativeai==0.8.3
openai==1.51.0

# === Embeddings & Vector Search ===
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
numpy==1.26.4

# === Data Pipeline ===
PyMuPDF==1.24.10
Pillow==10.4.0
pandas==2.2.3

# === Audio & Multimodal ===
elevenlabs==1.9.0
aiohttp==3.10.5
aiofiles==24.1.0

# === Fine-tuning ===
transformers==4.44.2
datasets==3.0.1
peft==0.13.0
trl==0.11.1
bitsandbytes==0.43.3
accelerate==0.34.2
torch>=2.1.0

# === Testing ===
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# === Utilities ===
tqdm==4.66.5
loguru==0.7.2
